correctly accounting for changes of stack and view
